[0.24] Move initOffsets functionality to a common package (#885)
* Move initOffsets functionality to a common package (#884)


* Move initOffsets functionality to a common package (#884)
